+++ src/pages/Office/MoveDetails/MoveDetails.jsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useMemo, useState } from 'react';
-import { Link, useNavigate, useParams } from 'react-router-dom';
+import { generatePath, Link, useNavigate, useParams } from 'react-router-dom';
 import { Alert, Grid, GridContainer } from '@trussworks/react-uswds';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { useQueryClient, useMutation } from '@tanstack/react-query';
@@ -15,6 +15,7 @@ import SERVICE_ITEM_STATUSES from 'constants/serviceItems';
 import { ADDRESS_UPDATE_STATUS, shipmentStatuses } from 'constants/shipments';
 import AllowancesList from 'components/Office/DefinitionLists/AllowancesList';
 import CustomerInfoList from 'components/Office/DefinitionLists/CustomerInfoList';
+import ButtonDropdown from 'components/ButtonDropdown/ButtonDropdown';
 import OrdersList from 'components/Office/DefinitionLists/OrdersList';
 import DetailsPanel from 'components/Office/DetailsPanel/DetailsPanel';
 import FinancialReviewButton from 'components/Office/FinancialReviewButton/FinancialReviewButton';
@@ -28,6 +29,7 @@ import LeftNavTag from 'components/LeftNavTag/LeftNavTag';
 import Restricted from 'components/Restricted/Restricted';
 import LoadingPlaceholder from 'shared/LoadingPlaceholder';
 import SomethingWentWrong from 'shared/SomethingWentWrong';
+import { SHIPMENT_OPTIONS_URL } from 'shared/constants';
 import { SIT_EXTENSION_STATUS } from 'constants/sitExtensions';
 import { ORDERS_TYPE } from 'constants/orders';
 import { permissionTypes } from 'constants/permissions';
@@ -154,6 +156,18 @@ const MoveDetails = ({
   const shipmentWithDestinationAddressChangeRequest = mtoShipments?.filter(
     (shipment) => shipment?.deliveryAddressUpdate?.status === ADDRESS_UPDATE_STATUS.REQUESTED && !shipment.deletedAt,
   );
+
+  const handleButtonDropdownChange = (e) => {
+    const selectedOption = e.target.value;
+
+    const addShipmentPath = `${generatePath(tooRoutes.SHIPMENT_ADD_PATH, {
+      moveCode,
+      shipmentType: selectedOption,
+    })}`;
+
+    navigate(addShipmentPath);
+  };
+
   useEffect(() => {
     const shipmentCount = shipmentWithDestinationAddressChangeRequest?.length || 0;
     if (setShipmentsWithDeliveryAddressUpdateRequestedCount)
@@ -363,6 +377,19 @@ const MoveDetails = ({
               </Grid>
             )}
           </Grid>
+          {!isMoveLocked && (
+            <Restricted to={permissionTypes.createTxoShipment}>
+              <ButtonDropdown data-testid="addShipmentButton" onChange={handleButtonDropdownChange}>
+                <option value="">Add a new shipment</option>
+                <option data-testid="hhgOption" value={SHIPMENT_OPTIONS_URL.HHG}>
+                  HHG
+                </option>
+                <option value={SHIPMENT_OPTIONS_URL.PPM}>PPM</option>
+                <option value={SHIPMENT_OPTIONS_URL.NTS}>NTS</option>
+                <option value={SHIPMENT_OPTIONS_URL.NTSrelease}>NTS-release</option>
+              </ButtonDropdown>
+            </Restricted>
+          )}
           {submittedShipments?.length > 0 && (
             <div className={styles.section} id="requested-shipments">
               <SubmittedRequestedShipments
